tcg: move locking for tb_invalidate_phys_page_range up
While we previously assumed an existing memory lock protected the page
look up in the MTTCG SoftMMU case the memory lock is provided by the
tb_lock. As a result we push the taking of this lock up the call tree.
This requires a slightly different entry for the SoftMMU and user-mode
cases from tb_invalidate_phys_range.

This also means user-mode breakpoint insertion needs to take two locks
but it hadn't taken any previously so this is an improvement.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>